hasql: cast int4 columns to bigint in two SELECTs
QueryCache.hs (hit_count) and Issues.hs (affected_requests,
affected_clients) decode into Haskell Int (= int8) but the underlying
columns are INTEGER (int4), so every dashboard fetch and every cache
lookup was throwing UnexpectedColumnTypeStatementError under hasql.
Cast on the SELECT rather than narrowing the record field — safer
because the same field types are used as Int across the codebase.

+++ src/Models/Apis/Issues.hs
@@ -445,7 +445,7 @@ selectIssues pid _typeM isAcknowledged isArchived limit offset timeRangeM sortM 
         -- FromRow/DecodeRow for Issue decodes the base row correctly.
         -- Prefer the stored severity (e.g. 'low' for silent drops); fall back to critical flag.
         COALESCE(NULLIF(i.severity, ''), CASE WHEN i.critical THEN 'critical' ELSE 'info' END),
-        i.affected_requests, i.affected_clients, NULL::double precision,
+        i.affected_requests::bigint, i.affected_clients::bigint, NULL::double precision,
         i.recommended_action, i.migration_complexity, i.issue_data, i.request_payloads, i.response_payloads,
         NULL::timestamp with time zone, NULL::bigint,
         i.target_hash, NULL::text, i.seq_num, i.parent_hash, i.is_framework,
